ai-email-assistant/src/requirements.txt
beautifulsoup4==4.13.3
duckduckgo_search==2025.4.4
google_api_python_client==2.160.0
google_auth_oauthlib==1.2.1
protobuf==6.30.2
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
pytz==2024.2
Requests==2.32.3
slack_sdk==3.35.0
